Rename CaptainContainer to CaptainVessel... And parse JSON objects via OJ callbacks
